Rename filenames from AbcDef.<ext> to abc-def.<ext>
We want to avoid casing in filenames due to git-limitations
