+++ .github/workflows/ci.yml
@@ -26,7 +26,7 @@ jobs:
             - run: npm ci
             - run: npm run test
 
-    svelte-check:
+    check:
         runs-on: ubuntu-latest
         steps:
             - uses: actions/checkout@v4
