+++ package.json
@@ -11,7 +11,7 @@
         "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
         "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
         "format": "prettier --write .",
-        "lint": "prettier --check . && eslint .",
+        "lint": "prettier --check . && eslint . && ./scripts/verify-filenames.sh",
         "test:unit": "vitest",
         "test": "npm run test:unit -- --run"
     },
