+++ src/lib/icons/asterisk-icon.svelte
