+++ src/lib/icons/calendar-icon.svelte
