+++ src/lib/icons/copy-icon.svelte
