+++ src/lib/icons/external-link-icon.svelte
