+++ src/lib/icons/graduation-icon.svelte
