+++ src/lib/icons/share-icon.svelte
