+++ src/lib/index.ts
@@ -1,8 +1,8 @@
-export { default as Input } from './Input.svelte';
-export { default as CronInput } from './CronInput.svelte';
-export { default as HumanReadable } from './HumanReadable.svelte';
-export { default as Schedules } from './Schedules.svelte';
-export { default as CopyButton } from './CopyButton.svelte';
-export { default as Card } from './Card.svelte';
-export { default as ExternalLink } from './ExternalLink.svelte';
+export { default as Input } from './input.svelte';
+export { default as CronInput } from './cron-input.svelte';
+export { default as HumanReadable } from './human-readable.svelte';
+export { default as Schedules } from './schedules.svelte';
+export { default as CopyButton } from './copy-button.svelte';
+export { default as Card } from './card.svelte';
+export { default as ExternalLink } from './external-link.svelte';
 export type { CronFieldType, CronFieldConfig, CronExpression, CronFieldChangeEvent } from './types';
